updated graph

+++ src/Resources/Graph.java
@@ -10,11 +10,11 @@ public class Graph {
 	private static class Edge{
 
 		Node destination;
-		int distance;
+		float distance;
 		int speedLimit;
 		float traffic;
 
-		public Edge(Node destination, int distance, int speedLimit, float traffic) {
+		public Edge(Node destination, float distance, int speedLimit, float traffic) {
 			this.destination = destination;
 			this.distance = distance;
 			this.speedLimit = speedLimit;
@@ -24,10 +24,10 @@ public class Graph {
 	
 	public static class Node{
 		String name;
-		int heuristic;
+		float heuristic;
 		boolean isGoal;
 		
-		public Node(String name, int heuristic, boolean isGoal) {
+		public Node(String name, float heuristic, boolean isGoal) {
 			this.name = name;
 			this.heuristic = heuristic;
 			this.isGoal = isGoal;
@@ -43,7 +43,7 @@ public class Graph {
 		map.put(s, new LinkedList<Edge>());
 	}
 	
-	public void addEdge(Node destination1, Node destination2, int distance, int speedLimit, float traffic) {
+	public void addEdge(Node destination1, Node destination2, float distance, int speedLimit, float traffic) {
 		if(!map.containsKey(destination1) || !map.containsKey(destination2)) {
 			throw new java.lang.Error("A destination has not been added");
 		}
